requirements.txt
# Core
numpy==1.24.4
opencv-python==4.7.0.72
scikit-image==0.21.0
matplotlib==3.7.1
pillow==10.1.0

# Feature Matching
scikit-learn==1.3.0

# Parallelization
joblib==1.3.2

# Optional: For evaluation and logging
seaborn==0.12.2
tqdm==4.66.1

# Jupyter (optional for notebooks)
notebook
